+++ forms/src/modules/formPipes/pipes/index.ts
@@ -1,0 +1,4 @@
+export * from './asFormArray/asFormArray.pipe';
+export * from './asFormControl/asFormControl.pipe';
+export * from './asFormGroup/asFormGroup.pipe';
+export * from './getControl/getControl.pipe';